migrate server, handlers, helpers to ts
