+++ server/lib/search.ts
